requirements.txt
aniso8601==6.0.0
Babel==2.6.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-RESTful==0.3.7
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
python-dotenv==0.10.1
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.15.2
